+++ forge/src/main/java/io/github/flemmli97/flan/FlanForge.java
@@ -7,6 +7,7 @@ import io.github.flemmli97.flan.forgeevent.ServerEvents;
 import io.github.flemmli97.flan.forgeevent.WorldEventsForge;
 import io.github.flemmli97.flan.scoreboard.ClaimCriterias;
 import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.common.Mod;
@@ -27,6 +28,7 @@ public class FlanForge {
         forge.addListener(ItemInteractEventsForge::useItem);
         forge.addListener(BlockInteractEventsForge::startBreakBlocks);
         forge.addListener(BlockInteractEventsForge::breakBlocks);
+        forge.addListener(EventPriority.HIGHEST, BlockInteractEventsForge::useBlocks);
         forge.addListener(EntityInteractEventsForge::attackEntity);
         forge.addListener(EntityInteractEventsForge::useAtEntity);
         forge.addListener(EntityInteractEventsForge::useEntity);