+++ fabric/src/main/java/io/github/flemmli97/flan/fabric/FlanFabric.java
@@ -1,7 +1,6 @@
 package io.github.flemmli97.flan.fabric;
 
 import io.github.flemmli97.flan.Flan;
-import io.github.flemmli97.flan.api.permission.ObjectToPermissionMap;
 import io.github.flemmli97.flan.commands.CommandClaim;
 import io.github.flemmli97.flan.config.ConfigHandler;
 import io.github.flemmli97.flan.event.BlockInteractEvents;
@@ -50,9 +49,8 @@ public class FlanFabric implements ModInitializer {
     }
 
     public static void serverLoad(MinecraftServer server) {
-        ConfigHandler.serverLoad(server);
-        ObjectToPermissionMap.reload(server);
         Flan.lockRegistry(server);
+        ConfigHandler.serverLoad(server);
     }
 
     public static void serverFinishLoad(MinecraftServer server) {