+++ common/src/main/java/io/github/flemmli97/flan/config/Config.java
@@ -22,6 +22,7 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Consumer;
+import java.util.stream.Stream;
 
 public class Config {
 
@@ -332,6 +333,23 @@ public class Config {
 
         Map<ClaimPermission, GlobalType> permMap = ConfigHandler.config.globalDefaultPerms.get(world.dimension().location().toString());
         return permMap == null ? GlobalType.NONE : permMap.getOrDefault(perm, GlobalType.NONE);
+    }
+
+    public Stream<Map.Entry<ClaimPermission, GlobalType>> getGloballyDefinedVals(ServerLevel world) {
+        Map<ClaimPermission, GlobalType> allMap = ConfigHandler.config.globalDefaultPerms.get("*");
+        if (allMap != null) {
+            world.getServer().getAllLevels().forEach(w -> {
+                Map<ClaimPermission, GlobalType> wMap = ConfigHandler.config.globalDefaultPerms.getOrDefault(w.dimension().location().toString(), new HashMap<>());
+                allMap.forEach((key, value) -> {
+                    if (!wMap.containsKey(key))
+                        wMap.put(key, value);
+                });
+                ConfigHandler.config.globalDefaultPerms.put(w.dimension().location().toString(), wMap);
+            });
+            ConfigHandler.config.globalDefaultPerms.remove("*");
+        }
+        Map<ClaimPermission, GlobalType> permMap = ConfigHandler.config.globalDefaultPerms.get(world.dimension().location().toString());
+        return permMap == null ? Stream.empty() : permMap.entrySet().stream().filter(e -> e.getValue().canModify());
     }
 
     public static <V, K> Map<V, K> createHashMap(Consumer<Map<V, K>> cons) {